# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HSD Auth Platform API
  description: |
    HSD Auth Platform - Authentication as a Service API
    
    Bu API, HSD projelerinde kimlik doğrulama işlemleri için kullanılır.
    
    ## Authentication
    API istekleri için JWT Bearer token veya API Key kullanılabilir.
    
    ## Rate Limiting
    - Varsayılan: 100 istek/dakika
    - Rate limit aşıldığında 429 Too Many Requests döner
  version: 1.0.0
  contact:
    name: HSD Auth Support
    email: support@hsdcore.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.auth.hsdcore.com
    description: Production server
  - url: https://api.staging.auth.hsdcore.com
    description: Staging server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /admin/realms/{realmId}/users:
    get:
      tags:
        - Admin
      summary: Realm kullanıcılarını listele
      operationId: listRealmUsers
      parameters:
        - $ref: '#/components/parameters/RealmId'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [active, disabled, pending]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Kullanıcı listesi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
    post:
      tags:
        - Admin
      summary: Kullanıcı oluştur
      operationId: createUser
      parameters:
        - $ref: '#/components/parameters/RealmId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Kullanıcı oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    RealmId:
      name: realmId
      in: path
      required: true
      schema:
        type: string
      description: Realm ID
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        status:
          type: string
          enum: [active, disabled, pending]
        profile:
          $ref: '#/components/schemas/Profile'
        roles:
          type: array
          items:
            type: string
        mfaEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Profile:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
        metadata:
          type: object
    ProfileInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
...
    CreateUserRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        profile:
          $ref: '#/components/schemas/ProfileInput'
        roles:
          type: array
          items:
            type: string
        emailVerified:
          type: boolean
        requirePasswordReset:
          type: boolean
...
    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer